added support for config safeMode + dropColumns

+++ src/db/queryBuilders/mysql/tableBuilder.ts
@@ -77,9 +77,11 @@ export class MySQLTableQueryBuilder {
         let alterStatements: string[] = [];
     
         // ✅ Handle `DROP COLUMN`
-        changes.dropColumns.forEach(columnName => {
-            alterStatements.push(`DROP COLUMN \`${columnName}\``);
-        });
+        if(databaseConfig?.deleteColumns) {
+            changes.dropColumns.forEach(columnName => {
+                alterStatements.push(`DROP COLUMN \`${columnName}\``);
+            });
+        }
     
         // ✅ Handle `RENAME COLUMN`
         changes.renameColumns.forEach(({ oldName, newName }) => {
